Make use of archive-writer abstraction to write gen/ directly to tar
