+++ scionlab/util/archive.py
@@ -1,0 +1,79 @@
+# Copyright 2018 ETH Zurich
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+import io
+import json
+import pathlib
+import tarfile
+import toml
+import yaml
+
+
+class BaseArchiveWriter:
+    def write_text(self, path, content):
+        raise NotImplementedError()
+
+    def write_json(self, path, content):
+        self.write_text(path, json.dumps(content, indent=2))
+
+    def write_toml(self, path, content):
+        self.write_text(path, toml.dumps(content))
+
+    def write_yaml(self, path, content):
+        self.write_text(path, yaml.dump(content, default_flow_style=False))
+
+    def write_config(self, path, config):
+        f = io.StringIO()
+        config.write(f)
+        self.write_text(path, f.getvalue())
+
+    def _normalize_path(self, path):
+        if isinstance(path, tuple):
+            return str(pathlib.PurePosixPath(*path))
+        else:
+            return path
+
+
+class FileArchiveWriter(BaseArchiveWriter):
+    def __init__(self, root):
+        self.root = root
+
+    def write_text(self, path, content):
+        filepath = pathlib.Path(self.root, self._normalize_path(path))
+        filepath.parent.mkdir(parents=True, exist_ok=True)
+        filepath.write_text(content)
+
+
+# Tape Archive Writer...
+class TarWriter(BaseArchiveWriter):
+    def __init__(self, tar):
+        self.tar = tar
+
+    def write_text(self, path, content):
+        path = self._normalize_path(path)
+        tar_add_textfile(self.tar, path, content)
+
+
+def tar_add_textfile(tar, path, content):
+    """
+    Helper for tarfile: add a text-file `name` with the given `content` to a tarfile `tar`.
+    :param TarFile tar: an open tarfile.TarFile
+    :param str path: name/path for the file in the tarfile
+    :param str content: file content
+    """
+    tarinfo = tarfile.TarInfo()
+    tarinfo.name = path
+    content_bytes = content.encode()
+    tarinfo.size = len(content_bytes)
+    tar.addfile(tarinfo, io.BytesIO(content_bytes))
